rest-api: introduce ResponseContent type for returning different content type payloads
